Remove BareField and InputTagMethodsTrait per review feedback
Hidden now extends BaseField directly with input-tag methods inlined.
Restore InputField.php and Theme.php to original implementations
(minus the Hidden-specific check in Theme for theme config support).
