+++ src/Field/Hidden.php
@@ -5,7 +5,8 @@ declare(strict_types=1);
 namespace Yiisoft\Form\Field;
 
 use InvalidArgumentException;
-use Yiisoft\Form\Field\Base\BareField;
+use Yiisoft\Form\Field\Base\BaseField;
+use Yiisoft\Form\Field\Base\InputData\InputDataWithCustomNameAndValueTrait;
 use Yiisoft\Html\Html;
 
 use function is_string;
@@ -17,8 +18,89 @@ use function is_string;
  * @link https://html.spec.whatwg.org/multipage/input.html#hidden-state-(type=hidden)
  * @link https://developer.mozilla.org/docs/Web/HTML/Element/input/hidden
  */
-final class Hidden extends BareField
+final class Hidden extends BaseField
 {
+    use InputDataWithCustomNameAndValueTrait;
+
+    protected bool $useContainer = false;
+
+    private ?string $inputId = null;
+    private bool $shouldSetInputId = true;
+    private array $inputAttributes = [];
+
+    public function inputId(?string $inputId): self
+    {
+        $new = clone $this;
+        $new->inputId = $inputId;
+        return $new;
+    }
+
+    public function shouldSetInputId(bool $value): self
+    {
+        $new = clone $this;
+        $new->shouldSetInputId = $value;
+        return $new;
+    }
+
+    public function inputAttributes(array $attributes): self
+    {
+        $new = clone $this;
+        $new->inputAttributes = $attributes;
+        return $new;
+    }
+
+    public function addInputAttributes(array $attributes): self
+    {
+        $new = clone $this;
+        $new->inputAttributes = array_merge($new->inputAttributes, $attributes);
+        return $new;
+    }
+
+    /**
+     * Replace input tag CSS classes with a new set of classes.
+     *
+     * @param string|null ...$class One or many CSS classes.
+     */
+    public function inputClass(?string ...$class): self
+    {
+        $new = clone $this;
+        $new->inputAttributes['class'] = array_filter($class, static fn($c) => $c !== null);
+        return $new;
+    }
+
+    /**
+     * Add one or more CSS classes to the input tag.
+     *
+     * @param string|null ...$class One or many CSS classes.
+     */
+    public function addInputClass(?string ...$class): self
+    {
+        $new = clone $this;
+        Html::addCssClass($new->inputAttributes, $class);
+        return $new;
+    }
+
+    public function form(?string $value): self
+    {
+        $new = clone $this;
+        $new->inputAttributes['form'] = $value;
+        return $new;
+    }
+
+    protected function getInputAttributes(): array
+    {
+        $attributes = $this->inputAttributes;
+
+        $this->prepareIdInInputAttributes($attributes);
+
+        return $attributes;
+    }
+
+    protected function generateContent(): ?string
+    {
+        return $this->generateInput();
+    }
+
     protected function generateInput(): string
     {
         $value = $this->getValue();
@@ -27,8 +109,21 @@ final class Hidden extends BareField
             throw new InvalidArgumentException('Hidden widget requires a string, numeric or null value.');
         }
 
-        $inputAttributes = $this->getInputAttributes();
+        return Html::hiddenInput(
+            $this->getName(),
+            $value,
+            $this->getInputAttributes(),
+        )->render();
+    }
 
-        return Html::hiddenInput($this->getName(), $value, $inputAttributes)->render();
+    private function prepareIdInInputAttributes(array &$attributes): void
+    {
+        if ($this->shouldSetInputId) {
+            if ($this->inputId !== null) {
+                $attributes['id'] = $this->inputId;
+            } elseif (!isset($attributes['id'])) {
+                $attributes['id'] = $this->getInputData()->getId();
+            }
+        }
     }
 }
